Normalize km values so the linear_regression converges faster and has less oscillation

+++ predict.py
@@ -5,7 +5,7 @@ import os
 
 from settings import T0_ENVAR_NAME, T1_ENVAR_NAME
 
-def estimate_price(t0: float, t1: float, mileage: int) -> float:
+def estimate_price(t0: float, t1: float, mileage: float) -> float:
     """
     Estimates the price of a car based on a linear regression model.
 
@@ -14,7 +14,7 @@ def estimate_price(t0: float, t1: float, mileage: int) -> float:
     :param t1: The slope (theta 1) of the linear regression model.
     :type t1: float
     :param mileage: The mileage of the car for which to estimate the price.
-    :type mileage: int
+    :type mileage: float
     :return: The estimated price of the car.
     :rtype: float
     """
